zsh: add the ability to locally override runcoms
Sometimes I don't want to append site-specific things in runcoms (as
already achieved by perhostenv, perhostrc, etc.); I want to override
entire files altogether.

For instance, on a server that logs me in to an AFS filesystem, I might
want to redirect my HOME to another filesystem that don't have the
limitations of AFS. In that case I don't want to set my env based on the
AFS HOME, so I would like to bypass the entire .zshenv if my desired
filesystem is mounted. This is achieved by
~/.config/zsh/zshenv_override like:

    if [[ -e /farmshare/user_data/zmwang ]]; then
        export HOME=/farmshare/user_data/zmwang
        source $HOME/.zshenv
        return 0
    else
        return 1
    fi

The return value of zshenv_override determines whether to continue
sourcing the regular zshenv: if it is zero, zshenv immediately
returns (in the above example, this happens if the target filesystem was
mounted and we were able to redirect HOME); otherwise, we continue to
evalute the rest of the regular zshenv (in the above example, this
happens if the target filesystem is not mounted, in which case we still
want to set the env).

The same mechanism applies to zprofile, zshrc, zlogin and zlogout, for
which we may have zprofile_override, zshrc_override, zlogin_override and
zlogout_override.