setup: mark some items as optional and skippable
Some items, e.g., pyenv/version, are not included in this directory and
thus might not be present. We mark those as optional and skippable,
print a warning for each such item that is not present, and report the
total number of skipped items at the end.

This commit also links pyenv/virtualenvwrapper_hooks as a whole instead
of linking hooks one by one (not sure what was in my mind when I did
that).