emacs/init.el: Unset env var GPG_TTY
When GPG_TTY is unset, pinentry-curses can't find the tty to spin up its
curses window, so it will fail automatically. In this case I'll have to
use my gpgauth function from outside Emacs to populate the cache of
gpg-agent.

This is much better than pinentry-curses competing with Emacs and
subsequently mess up everything. (In that case, one out of every two
keystrokes seem to go to pinentry-curses and the other go to the
controlling tty, so it's very hard even just to kill pinentry-curses).